Refactored conditioners into their own files
